+++ libraries/bio/CHANGELOG.md
@@ -26,6 +26,7 @@
 * Fix monomer lib handler for PolyTool
 * Fix splitter with separator for quoted gaps
 * Fix default value for pdbTag prop of Biostructure
+* Fix biostructure data provider to return JSON string
 
 ## 5.38.0 (2023-09-06)